pyproject.toml
[project]
name = "clubchess"
version = "0.1.0"
description = ""
authors = [
    {name = "Natalia",email = "lasinforosa@gmail.com"}
]
requires-python = ">=3.14"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "python-chess (>=1.999,<2.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "django-taggit (>=6.1.0,<7.0.0)",
    "dj-database-url (>=3.1.2,<4.0.0)",
    "dotenv (>=0.9.9,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
